+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Bump `piggieback` 0.6.0 → 0.6.1 (test dependency).
 * Bump `orchard` 0.39.0 → 0.41.0.
 * Bump `error_prone_annotations` 2.20.0 → 2.49.0 (pedantic dependency).
+* Bump `clojure` (in matrix's `:1.12` cell) 1.12.0 → 1.12.4.
 
 ## 3.12.0 (2026-05-10)